FactionPointsBonus
Changes to the faction points system:
From now on, if you're part of a faction and get a rival kill (for example, an Armada member killing a Legion member or vice versa), the points gained from that kill will be increased by a multiplier set in the Server configuration.

Additionally, the system now includes a bonus for killing an enemy of a higher level than yours. Previously, you would only lose points for killing someone of lower level, and if you killed a rival of higher level, the reward was capped at 10 points.
With this change, it can now scale up to 20 points — meaning that killing an enemy who is 10 or more levels above you can grant you up to 20 faction points.