Refactor equipment skin selection logic
Updated DoNavega and DoMontar to improve handling of equipped item skins. Now, equipped skin indices for armor, helmet, shield, and weapon are checked and applied correctly, and default animations are set when no equipment is present.
Prior to this change, when any user unmounted or got off the ship their skin would've been unequipped (visually, not logically) making the person re-equip the skin if they wanted to show it again