Removed exit sub
Deleted an unnecessary 'Exit Sub' line after a conditional block.
Probably a bug introduced in #995.
This fixes a bug where equipped weapons dont show correctly as "equipped" in the client, resulting in bugs where classes that use "consumable" or "interactable" weapons like the bow get their weapon UN-EQUIPPED when they are trying to use it